+++ src/CFDIReader/PostValidations/Validators/Fechas.php
@@ -6,13 +6,21 @@ use CFDIReader\PostValidations\Issues;
 
 class Fechas extends AbstractValidator
 {
+    /**
+     * Store the tolerance in seconds when comparing dates. Default to 1 minute.
+     * This prevents the problem of time sync between the emitter and the PAC.
+     *
+     * @var int
+     */
+    private $delta = 60;
+
     public function validate(CFDIReader $cfdi, Issues $issues)
     {
         // setup the AbstractValidator Helper class
         $this->setup($cfdi, $issues);
         // do the validation process
         $document = strtotime($cfdi->attribute('fecha'));
-        if ($document > $this->getCurrentDate()) {
+        if ($document > $this->getCurrentDate() + $this->getDelta()) {
             $this->errors->add('La fecha del documento es mayor a la fecha actual');
         }
         if ($cfdi->hasTimbreFiscalDigital()) {
@@ -21,14 +29,24 @@ class Fechas extends AbstractValidator
                 $this->errors->add('Existe un timbre pero no tiene fecha de timbrado');
             } else {
                 $timbrado = strtotime($timbrado);
-                if ($document > $timbrado) {
+                if ($document > $timbrado + $this->getDelta()) {
                     $this->errors->add('La fecha del documento es mayor a la fecha del timbrado');
                 }
-                if ($timbrado - $document > 259200) {
+                if ($timbrado > $document + 259200 + $this->getDelta()) {
                     $this->errors->add('La fecha fecha del timbrado excedió las 72 horas de la fecha del documento');
                 }
             }
         }
+    }
+
+    public function getDelta(): int
+    {
+        return $this->delta;
+    }
+
+    public function setDelta(int $delta)
+    {
+        $this->delta = $delta;
     }
 
     /**